LPS-205650 Extract ScreenNavigationEntry methods from ServerSystemPropertiesScreenNavigationCategory to ServerSystemPropertiesScreenNavigationEntry and extends ServerSystemPropertiesScreenNavigationCategory in it
