+++ modules/apps/server/server-admin-web/src/main/java/com/liferay/server/admin/web/internal/frontend/taglib/servlet/taglib/ServerSystemPropertiesScreenNavigationCategory.java
@@ -6,17 +6,10 @@
 package com.liferay.server.admin.web.internal.frontend.taglib.servlet.taglib;
 
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.server.admin.web.internal.constants.ServerAdminNavigationEntryConstants;
 
-import java.io.IOException;
-
 import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -25,14 +18,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Albert Lee
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=10",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=10",
+	service = ScreenNavigationCategory.class
 )
 public class ServerSystemPropertiesScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<Object> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -41,13 +31,8 @@ public class ServerSystemPropertiesScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return ServerAdminNavigationEntryConstants.ENTRY_KEY_SYSTEM_PROPERTIES;
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, "system-properties");
+		return language.get(locale, "system-properties");
 	}
 
 	@Override
@@ -56,21 +41,7 @@ public class ServerSystemPropertiesScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_PROPERTIES;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/view_system_properties.jsp");
-	}
-
 	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
+	protected Language language;
 
 }
